+++ modules/repost.ts
@@ -1,23 +1,28 @@
-import { Events, PermissionsBitField } from 'discord.js';
+import { Events, Message, TextChannel, PermissionsBitField } from 'discord.js';
 
-export default {
-    name: Events.MessageCreate,
-    once: false,
-    async execute(message) {
+export default class Repost implements botModule {
+    name = "Repost";
+    version = "1.0";
+    type = Events.MessageCreate;
+    once = false;
+    async execute(message: Message) {
         if (message.author.bot) return;
 
         const args = message.content.split(/\s+/);
-        const command = args.shift().toLowerCase();
+        const command = args.shift()?.toLowerCase();
 
         if (command === "!repost") {
+            if (message.member == null) return;
             if (message.member.permissions.has(PermissionsBitField.Flags.ManageMessages)) {
                 const content = message.content.slice(command.length).trim();
                 if (content.length > 0) {
+                    if (!(message.channel instanceof TextChannel)) return;
                     await message.channel.send(content);
                 }
 
                 if (message.attachments.size > 0) {
-                    await message.channel.send({ files: [`${message.attachments.first().url}`] })
+                    if (!(message.channel instanceof TextChannel)) return;
+                    await message.channel.send({ files: [`${message.attachments.first()?.url}`] })
                 }
 
                 message.delete();
@@ -25,6 +30,5 @@ export default {
                 await message.reply("You don't have permission to use this command.");
             }
         }
-
-    },
+    }
 };
